fixed rotation for adding entities/small bugs for integer rounding error related things

+++ Bawk/placeablesuperobject.h
@@ -32,8 +32,8 @@ public:
     
     // --- SuperObject ---
     int get_chunk(block_type to_arr[CX][CY][CZ], int x, int y, int z) override;
-    void handle_block_addition(float x, float y, float z, block_type type) override;
-    void handle_block_removal(float x, float y, float z, block_type type) override;
+    void handle_block_addition(int x, int y, int z, block_type type) override;
+    void handle_block_removal(int x, int y, int z, block_type type) override;
     void update_chunks(fvec3* start_pos) override;
     
     // --- PlaceableObject ---
